Fix run time errors
